Added junit to the build system as part of ticket #28.
This added the following ant targets:

   * junit-tests: Runs the tests (everything under ${junit-test-src-dir}
that matches the pattern **/*Test.java).  This target should "just work" --
building whatever is necessary and then running the tests.
   * build-junit-tests: Builds the test source (at the moment, it just
shoves the classes into the ${build-dir})

Results are recorded to a file in `BeanShell/tests/junit-reports`, and that
directory is cleared on `ant clean`.  A summary report is also printed at
the end of the ant output.

git-svn-id: http://beanshell2.googlecode.com/svn@35 8af74f78-9a2f-0410-974c-5db699c9257e